+++ packages/source-client/src/implementations/pg/PGEditor.ts
@@ -21,6 +21,14 @@ export class PGEditor<T> extends PGReader<T> implements Editable<T> {
             this.identificationValues(),
         )
     }
+    public async isRestorable(): Promise<boolean> {
+        const client = await this.provider.getPG()
+        const output = await client.query(
+            `SELECT ${this.identificationColumns()} FROM ${this.table} WHERE ${this.identification()}`,
+            this.identificationValues(),
+        )
+        return (output.rowCount ?? 0) >= 1
+    }
     public async put(value: T) {
         const client = await this.provider.getPG()
         if (await this.exists()) {
@@ -36,6 +44,17 @@ export class PGEditor<T> extends PGReader<T> implements Editable<T> {
                 this.insertValues(value),
             )
         }
+    }
+    public async restore() {
+        if (await this.isRestorable()) {
+            const client = await this.provider.getPG()
+            await client.query(
+                `UPDATE ${this.table} SET disabled=0::bit WHERE ${this.identification()}`,
+                this.identificationValues(),
+            )
+            return this.get()
+        }
+        throw new Error("Cannot restore.")
     }
     protected insertColumns() {
         return this.fields