+++ packages/source-client/src/implementations/s3/S3Deletor.ts
@@ -1,8 +1,10 @@
 import { GetObjectOutput } from "@aws-sdk/client-s3"
 import { Deletable } from "../../interfaces/Deletable"
 import { S3ClientProvider } from "../../interfaces/S3ClientProvider"
+import { copyFromTrash } from "./methods/copyFromTrash"
 import { copyToTrash } from "./methods/copyToTrash"
 import { deleteObject } from "./methods/deleteObject"
+import { exists } from "./methods/exists"
 import { S3Reader } from "./S3Reader"
 export class S3Deletor<T> extends S3Reader<T> implements Deletable<T> {
     constructor(
@@ -18,6 +20,20 @@ export class S3Deletor<T> extends S3Reader<T> implements Deletable<T> {
             await this.copyToTrash()
             await deleteObject(this.provider.getS3(), this.bucket, this.key)
         }
+    }
+    public async isRestorable(): Promise<boolean> {
+        return exists(this.provider.getS3(), this.bucket, `trash/${this.key}`)
+    }
+    public async restore() {
+        if (await this.isRestorable()) {
+            await this.copyFromTrash()
+            await deleteObject(this.provider.getS3(), this.bucket, `trash/${this.key}`)
+            return this.get()
+        }
+        throw new Error("Cannot restore")
+    }
+    protected async copyFromTrash() {
+        await copyFromTrash(this.provider.getS3(), this.bucket, this.key)
     }
     protected async copyToTrash() {
         await copyToTrash(this.provider.getS3(), this.bucket, this.key)