+++ packages/source-client/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   New methods to Deletor clients: `isRestorable()`, `restore()`.
+
 ### Changed
 
 ### Deprecated